# ProjectManagement_Backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is Project Management API documentation",
        "title": "Project Management APIs",
        "contact": {},
        "version": "1.0"
    },
    "host": "projectmanagement-backend-tysm.onrender.com",
    "paths": {
...
        "/api/v1/work": {
            "get": {
                "description": "Get all works",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "work"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Get all works",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Work"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new work",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "work"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Create a new work",
                "parameters": [
                    {
                        "description": "Project",
                        "name": "work",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Work"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Work"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/work/{id}": {
            "get": {
                "description": "Get work by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "work"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Get work by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Work ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Work"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Update work by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "work"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Update work by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Work ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Work",
                        "name": "work",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Work"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Work"
                        }
                    },
                    "400": {
                        "description": "Invalid ID or Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete work by ID",
                "tags": [
                    "work"
                ],
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Delete work by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Work ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}